Import helper functions from utils
